% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-curtains.R
\name{makeCurtains}
\alias{makeCurtains}
\title{Make my curtains app}
\usage{
makeCurtains(port = 4455L)
}
\arguments{
\item{port}{Numeric port number.}
}
\value{
Shiny application.
}
\description{
Open sesame!
}
\examples{
\dontrun{
makeCurtains()
}
}
